split index js code + change export structure + changed banner
