+++ src/tests/basic.test.js
@@ -1,4 +1,4 @@
-import { JSONGrid } from "..";
+import JSONGrid from "..";
 
 describe("JSONGrid", () => {
   it("is truthy", () => {
